Multiple perf improvements, see details in each commits.

A global benchmark before and after to compare the improvements


| Method                          | Before (avg of 5 runs) | After  (avg of 5 runs) | Gain         | Gain % |
|---------------------------------|------------------------|------------------------|--------------|--------|
| BigDecimal.add                  | 5.361666483s           | 3.626009725s           | 1.735656758s | 32.4%  |
| BigDecimal.divide               | 17.563635691s          | 13.214709058s          | 4.348926633s | 24.8%  |
| BigDecimal.hashCode             | 2.090754216s           | 1.885830800s           | 0.204923416s | 9.8%   |
| BigDecimal.isZero               | 168.348158ms           | 120.002383ms           | 48.345775ms  | 28.7%  |
| BigDecimal.multiply             | 4.306248025s           | 3.895518808s           | 0.410729217s | 9.5%   |
| BigDecimal.parseStringWithMode  | 6.032730558s           | 2.795508533s           | 3.237222025s | 53.7%  |
| BigDecimal.pow                  | 9.263568549s           | 8.316417325s           | 0.947151224s | 10.2%  |
| BigDecimal.roundToDigitPosition | 3.931222933s           | 3.151020300s           | 0.780202633s | 19.8%  |
| BigDecimal.subtract             | 4.932343592s           | 3.363048383s           | 1.569295209s | 31.8%  |
| BigDecimal.toPlainString        | 3.240023633s           | 3.170939741s           | 0.069083892s | 2.1% (luck)   |
| BigDecimal.toStringExpanded     | 3.181326449s           | 3.124509074s           | 0.056817375s | 1.8% (luck)  |
| BigInteger.TEN.pow              | 201.281508ms           | 120.361291ms           | 80.920217ms  | 40.2%  |
| BigInteger.isZero               | 117.322116ms           | 33.185292ms            | 84.136824ms  | 71.7%  |
| BigInteger.pow                  | 634.312616ms           | 519.652508ms           | 114.660108ms | 18.1%  |

<details>
<summary>All details are available below</summary>



PERF RESEARCH

ORIGINAL PERF BASED ON 384a12c9365a957e3ae7df372a7cd476f661a088

--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Preparing for BigDecimal.parseStringWithMode (warmup)
Benchmarking for BigDecimal.parseStringWithMode
Benchmark 0 DONE for BigDecimal.parseStringWithMode in 6.129106709s
Benchmark 1 DONE for BigDecimal.parseStringWithMode in 6.069024791s
Benchmark 2 DONE for BigDecimal.parseStringWithMode in 6.033444667s
Benchmark 3 DONE for BigDecimal.parseStringWithMode in 5.967124125s
Benchmark 4 DONE for BigDecimal.parseStringWithMode in 5.964952500s
Preparing for BigDecimal.add (warmup)
Benchmarking for BigDecimal.add
Benchmark 0 DONE for BigDecimal.add in 5.433343625s
Benchmark 1 DONE for BigDecimal.add in 5.355922291s
Benchmark 2 DONE for BigDecimal.add in 5.331163333s
Benchmark 3 DONE for BigDecimal.add in 5.342635542s
Benchmark 4 DONE for BigDecimal.add in 5.345267625s
Preparing for BigDecimal.subtract (warmup)
Benchmarking for BigDecimal.subtract
Benchmark 0 DONE for BigDecimal.subtract in 4.945782625s
Benchmark 1 DONE for BigDecimal.subtract in 4.903031334s
Benchmark 2 DONE for BigDecimal.subtract in 4.908852209s
Benchmark 3 DONE for BigDecimal.subtract in 4.977049625s
Benchmark 4 DONE for BigDecimal.subtract in 4.927002167s
Preparing for BigDecimal.multiply (warmup)
Benchmarking for BigDecimal.multiply
Benchmark 0 DONE for BigDecimal.multiply in 4.395549250s
Benchmark 1 DONE for BigDecimal.multiply in 4.337605s
Benchmark 2 DONE for BigDecimal.multiply in 4.286808708s
Benchmark 3 DONE for BigDecimal.multiply in 4.254930542s
Benchmark 4 DONE for BigDecimal.multiply in 4.256346625s
Preparing for BigDecimal.divide (warmup)
Benchmarking for BigDecimal.divide
Benchmark 0 DONE for BigDecimal.divide in 17.197246625s
Benchmark 1 DONE for BigDecimal.divide in 17.178957208s
Benchmark 2 DONE for BigDecimal.divide in 17.370274125s
Benchmark 3 DONE for BigDecimal.divide in 18.449112917s
Benchmark 4 DONE for BigDecimal.divide in 17.622587583s
Preparing for BigDecimal.hashCode (warmup)
Benchmarking for BigDecimal.hashCode
Benchmark 0 DONE for BigDecimal.hashCode in 2.099849666s
Benchmark 1 DONE for BigDecimal.hashCode in 2.081946875s
Benchmark 2 DONE for BigDecimal.hashCode in 2.087968833s
Benchmark 3 DONE for BigDecimal.hashCode in 2.083653583s
Benchmark 4 DONE for BigDecimal.hashCode in 2.100352125s
Preparing for BigDecimal.toStringExpanded (warmup)
Benchmarking for BigDecimal.toStringExpanded
Benchmark 0 DONE for BigDecimal.toStringExpanded in 3.173884916s
Benchmark 1 DONE for BigDecimal.toStringExpanded in 3.218394625s
Benchmark 2 DONE for BigDecimal.toStringExpanded in 3.180223375s
Benchmark 3 DONE for BigDecimal.toStringExpanded in 3.170393708s
Benchmark 4 DONE for BigDecimal.toStringExpanded in 3.163735625s
Preparing for BigDecimal.toPlainString (warmup)
Benchmarking for BigDecimal.toPlainString
Benchmark 0 DONE for BigDecimal.toPlainString in 3.219728875s
Benchmark 1 DONE for BigDecimal.toPlainString in 3.187645167s
Benchmark 2 DONE for BigDecimal.toPlainString in 3.350294417s
Benchmark 3 DONE for BigDecimal.toPlainString in 3.232167166s
Benchmark 4 DONE for BigDecimal.toPlainString in 3.210282542s
Preparing for BigDecimal.roundToDigitPosition (warmup)
Benchmarking for BigDecimal.roundToDigitPosition
Benchmark 0 DONE for BigDecimal.roundToDigitPosition in 3.929448208s
Benchmark 1 DONE for BigDecimal.roundToDigitPosition in 3.919461792s
Benchmark 2 DONE for BigDecimal.roundToDigitPosition in 3.954164375s
Benchmark 3 DONE for BigDecimal.roundToDigitPosition in 3.931849625s
Benchmark 4 DONE for BigDecimal.roundToDigitPosition in 3.921190667s
Preparing for BigDecimal.pow (warmup)
Benchmarking for BigDecimal.pow
Benchmark 0 DONE for BigDecimal.pow in 9.293656458s
Benchmark 1 DONE for BigDecimal.pow in 9.614776375s
Benchmark 2 DONE for BigDecimal.pow in 9.107468916s
Benchmark 3 DONE for BigDecimal.pow in 9.118559875s
Benchmark 4 DONE for BigDecimal.pow in 9.183381125s
Preparing for BigDecimal.isZero (warmup)
Benchmarking for BigDecimal.isZero
Benchmark 0 DONE for BigDecimal.isZero in 315.393791ms
Benchmark 1 DONE for BigDecimal.isZero in 255.624959ms
Benchmark 2 DONE for BigDecimal.isZero in 121.243584ms
Benchmark 3 DONE for BigDecimal.isZero in 88.027583ms
Benchmark 4 DONE for BigDecimal.isZero in 61.450875ms
Preparing for BigInteger.pow (warmup)
Benchmarking for BigInteger.pow
Benchmark 0 DONE for BigInteger.pow in 634.632917ms
Benchmark 1 DONE for BigInteger.pow in 630.169709ms
Benchmark 2 DONE for BigInteger.pow in 629.683250ms
Benchmark 3 DONE for BigInteger.pow in 633.617583ms
Benchmark 4 DONE for BigInteger.pow in 643.459625ms
Preparing for BigInteger.TEN.pow (warmup)
Benchmarking for BigInteger.TEN.pow
Benchmark 0 DONE for BigInteger.TEN.pow in 201.571416ms
Benchmark 1 DONE for BigInteger.TEN.pow in 200.507542ms
Benchmark 2 DONE for BigInteger.TEN.pow in 201.433792ms
Benchmark 3 DONE for BigInteger.TEN.pow in 201.285334ms
Benchmark 4 DONE for BigInteger.TEN.pow in 201.609459ms
Preparing for BigInteger.isZero (warmup)
Benchmarking for BigInteger.isZero
Benchmark 0 DONE for BigInteger.isZero in 124.876250ms
Benchmark 1 DONE for BigInteger.isZero in 57.415542ms
Benchmark 2 DONE for BigInteger.isZero in 40.308125ms
Benchmark 3 DONE for BigInteger.isZero in 40.634041ms
Benchmark 4 DONE for BigInteger.isZero in 323.376625ms
Benchmark report:

| Method |  run 1 | run 2 | run 3 | run 4 | run 5  | Average |
|--------|-------|------|------|------|------ |---------|
| BigDecimal.add |  5.433343625s | 5.355922291s | 5.331163333s | 5.342635542s | 5.345267625s  | 5.361666483s |
| BigDecimal.divide |  17.197246625s | 17.178957208s | 17.370274125s | 18.449112917s | 17.622587583s  | 17.563635691s |
| BigDecimal.hashCode |  2.099849666s | 2.081946875s | 2.087968833s | 2.083653583s | 2.100352125s  | 2.090754216s |
| BigDecimal.isZero |  315.393791ms | 255.624959ms | 121.243584ms | 88.027583ms | 61.450875ms  | 168.348158ms |
| BigDecimal.multiply |  4.395549250s | 4.337605s | 4.286808708s | 4.254930542s | 4.256346625s  | 4.306248025s |
| BigDecimal.parseStringWithMode |  6.129106709s | 6.069024791s | 6.033444667s | 5.967124125s | 5.964952500s  | 6.032730558s |
| BigDecimal.pow |  9.293656458s | 9.614776375s | 9.107468916s | 9.118559875s | 9.183381125s  | 9.263568549s |
| BigDecimal.roundToDigitPosition |  3.929448208s | 3.919461792s | 3.954164375s | 3.931849625s | 3.921190667s  | 3.931222933s |
| BigDecimal.subtract |  4.945782625s | 4.903031334s | 4.908852209s | 4.977049625s | 4.927002167s  | 4.932343592s |
| BigDecimal.toPlainString |  3.219728875s | 3.187645167s | 3.350294417s | 3.232167166s | 3.210282542s  | 3.240023633s |
| BigDecimal.toStringExpanded |  3.173884916s | 3.218394625s | 3.180223375s | 3.170393708s | 3.163735625s  | 3.181326449s |
| BigInteger.TEN.pow |  201.571416ms | 200.507542ms | 201.433792ms | 201.285334ms | 201.609459ms  | 201.281508ms |
| BigInteger.isZero |  124.876250ms | 57.415542ms | 40.308125ms | 40.634041ms | 323.376625ms  | 117.322116ms |
| BigInteger.pow |  634.632917ms | 630.169709ms | 629.683250ms | 633.617583ms | 643.459625ms  | 634.312616ms |

-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- after 1st commit e1028814fb6156579aa9a0152d128c53e6b31a92



Preparing for parseStringWithMode (warmup)
Benchmarking for parseStringWithMode
Benchmark 0 DONE for parseStringWithMode in 5.871168875s
Benchmark 1 DONE for parseStringWithMode in 5.817717417s
Benchmark 2 DONE for parseStringWithMode in 6.086022375s
Benchmark 3 DONE for parseStringWithMode in 5.823671166s
Benchmark 4 DONE for parseStringWithMode in 5.786276500s
Preparing for add (warmup)
Benchmarking for add
Benchmark 0 DONE for add in 5.385408584s
Benchmark 1 DONE for add in 5.359253500s
Benchmark 2 DONE for add in 5.217283916s
Benchmark 3 DONE for add in 5.213005542s
Benchmark 4 DONE for add in 5.226768917s
Preparing for subtract (warmup)
Benchmarking for subtract
Benchmark 0 DONE for subtract in 4.833749708s
Benchmark 1 DONE for subtract in 4.737383542s
Benchmark 2 DONE for subtract in 4.668813250s
Benchmark 3 DONE for subtract in 4.795859792s
Benchmark 4 DONE for subtract in 4.706975333s
Preparing for multiply (warmup)
Benchmarking for multiply
Benchmark 0 DONE for multiply in 4.289741750s
Benchmark 1 DONE for multiply in 4.164642250s
Benchmark 2 DONE for multiply in 4.114723625s
Benchmark 3 DONE for multiply in 4.065344458s
Benchmark 4 DONE for multiply in 4.604802333s
Preparing for divide (warmup)
Benchmarking for divide
Benchmark 0 DONE for divide in 16.007358750s
Benchmark 1 DONE for divide in 16.290671208s
Benchmark 2 DONE for divide in 16.348688791s
Benchmark 3 DONE for divide in 16.214174333s
Benchmark 4 DONE for divide in 16.129669125s
Preparing for hashCode (warmup)
Benchmarking for hashCode
Benchmark 0 DONE for hashCode in 2.248416375s
Benchmark 1 DONE for hashCode in 2.144754208s
Benchmark 2 DONE for hashCode in 2.121003708s
Benchmark 3 DONE for hashCode in 2.110207208s
Benchmark 4 DONE for hashCode in 2.115298583s
Preparing for toStringExpanded (warmup)
Benchmarking for toStringExpanded
Benchmark 0 DONE for toStringExpanded in 3.154393917s
Benchmark 1 DONE for toStringExpanded in 3.135770167s
Benchmark 2 DONE for toStringExpanded in 3.125461583s
Benchmark 3 DONE for toStringExpanded in 3.122853042s
Benchmark 4 DONE for toStringExpanded in 3.127719417s
Preparing for toPlainString (warmup)
Benchmarking for toPlainString
Benchmark 0 DONE for toPlainString in 3.221781750s
Benchmark 1 DONE for toPlainString in 3.279738458s
Benchmark 2 DONE for toPlainString in 3.269062042s
Benchmark 3 DONE for toPlainString in 3.242374834s
Benchmark 4 DONE for toPlainString in 3.361823833s
Preparing for roundToDigitPosition (warmup)
Benchmarking for roundToDigitPosition
Benchmark 0 DONE for roundToDigitPosition in 4.082506875s
Benchmark 1 DONE for roundToDigitPosition in 4.087423667s
Benchmark 2 DONE for roundToDigitPosition in 4.323030167s
Benchmark 3 DONE for roundToDigitPosition in 4.081716666s
Benchmark 4 DONE for roundToDigitPosition in 4.076284833s
Preparing for pow (warmup)
Benchmarking for pow
Benchmark 0 DONE for pow in 8.825030750s
Benchmark 1 DONE for pow in 8.786915709s
Benchmark 2 DONE for pow in 8.720024125s
Benchmark 3 DONE for pow in 8.677275042s
Benchmark 4 DONE for pow in 8.600209875s
Benchmark report:

| Method |  run 1 | run 2 | run 3 | run 4 | run 5  | Average |
|--------|-------|------|------|------|------ |---------|
| add |  5.385408584s | 5.359253500s | 5.217283916s | 5.213005542s | 5.226768917s  | 5.280344091s |
| divide |  16.007358750s | 16.290671208s | 16.348688791s | 16.214174333s | 16.129669125s  | 16.198112441s |
| hashCode |  2.248416375s | 2.144754208s | 2.121003708s | 2.110207208s | 2.115298583s  | 2.147936016s |
| multiply |  4.289741750s | 4.164642250s | 4.114723625s | 4.065344458s | 4.604802333s  | 4.247850883s |
| parseStringWithMode |  5.871168875s | 5.817717417s | 6.086022375s | 5.823671166s | 5.786276500s  | 5.876971266s |
| pow |  8.825030750s | 8.786915709s | 8.720024125s | 8.677275042s | 8.600209875s  | 8.721891100s |
| roundToDigitPosition |  4.082506875s | 4.087423667s | 4.323030167s | 4.081716666s | 4.076284833s  | 4.130192441s |
| subtract |  4.833749708s | 4.737383542s | 4.668813250s | 4.795859792s | 4.706975333s  | 4.748556325s |
| toPlainString |  3.221781750s | 3.279738458s | 3.269062042s | 3.242374834s | 3.361823833s  | 3.274956183s |
| toStringExpanded |  3.154393917s | 3.135770167s | 3.125461583s | 3.122853042s | 3.127719417s  | 3.133239625s |


-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- after 3rd commit 025a77465a7da03bc2d4f9df8024f1539e57db34


Preparing for parseStringWithMode (warmup)
Benchmarking for parseStringWithMode
Benchmark 0 DONE for parseStringWithMode in 5.962909500s
Benchmark 1 DONE for parseStringWithMode in 5.879412792s
Benchmark 2 DONE for parseStringWithMode in 5.981781583s
Benchmark 3 DONE for parseStringWithMode in 5.899648292s
Benchmark 4 DONE for parseStringWithMode in 5.923697917s
Preparing for add (warmup)
Benchmarking for add
Benchmark 0 DONE for add in 5.435422292s
Benchmark 1 DONE for add in 5.492117917s
Benchmark 2 DONE for add in 5.418271250s
Benchmark 3 DONE for add in 5.384207958s
Benchmark 4 DONE for add in 5.309335208s
Preparing for subtract (warmup)
Benchmarking for subtract
Benchmark 0 DONE for subtract in 5.068563625s
Benchmark 1 DONE for subtract in 4.885288833s
Benchmark 2 DONE for subtract in 4.796286s
Benchmark 3 DONE for subtract in 4.917022458s
Benchmark 4 DONE for subtract in 4.777188125s
Preparing for multiply (warmup)
Benchmarking for multiply
Benchmark 0 DONE for multiply in 4.233069167s
Benchmark 1 DONE for multiply in 4.212958458s
Benchmark 2 DONE for multiply in 4.212513125s
Benchmark 3 DONE for multiply in 4.230453625s
Benchmark 4 DONE for multiply in 4.185100459s
Preparing for divide (warmup)
Benchmarking for divide
Benchmark 0 DONE for divide in 16.158979458s
Benchmark 1 DONE for divide in 16.077749s
Benchmark 2 DONE for divide in 16.256035583s
Benchmark 3 DONE for divide in 16.137719667s
Benchmark 4 DONE for divide in 16.035147041s
Preparing for hashCode (warmup)
Benchmarking for hashCode
Benchmark 0 DONE for hashCode in 2.037175958s
Benchmark 1 DONE for hashCode in 2.095293416s
Benchmark 2 DONE for hashCode in 2.011758750s
Benchmark 3 DONE for hashCode in 1.999053083s
Benchmark 4 DONE for hashCode in 2.009785958s
Preparing for toStringExpanded (warmup)
Benchmarking for toStringExpanded
Benchmark 0 DONE for toStringExpanded in 3.206883875s
Benchmark 1 DONE for toStringExpanded in 3.205201916s
Benchmark 2 DONE for toStringExpanded in 3.372421625s
Benchmark 3 DONE for toStringExpanded in 3.204819125s
Benchmark 4 DONE for toStringExpanded in 3.175107s
Preparing for toPlainString (warmup)
Benchmarking for toPlainString
Benchmark 0 DONE for toPlainString in 3.276761500s
Benchmark 1 DONE for toPlainString in 3.318301750s
Benchmark 2 DONE for toPlainString in 3.604505417s
Benchmark 3 DONE for toPlainString in 3.422065917s
Benchmark 4 DONE for toPlainString in 3.557957958s
Preparing for roundToDigitPosition (warmup)
Benchmarking for roundToDigitPosition
Benchmark 0 DONE for roundToDigitPosition in 4.028694209s
Benchmark 1 DONE for roundToDigitPosition in 4.146693625s
Benchmark 2 DONE for roundToDigitPosition in 3.875029625s
Benchmark 3 DONE for roundToDigitPosition in 3.903158042s
Benchmark 4 DONE for roundToDigitPosition in 3.845113625s
Preparing for pow (warmup)
Benchmarking for pow
Benchmark 0 DONE for pow in 9.133553458s
Benchmark 1 DONE for pow in 9.139799458s
Benchmark 2 DONE for pow in 8.944042083s
Benchmark 3 DONE for pow in 9.104886416s
Benchmark 4 DONE for pow in 9.017410458s
Benchmark report:

| Method |  run 1 | run 2 | run 3 | run 4 | run 5  | Average |
|--------|-------|------|------|------|------ |---------|
| add |  5.435422292s | 5.492117917s | 5.418271250s | 5.384207958s | 5.309335208s  | 5.407870925s |
| divide |  16.158979458s | 16.077749s | 16.256035583s | 16.137719667s | 16.035147041s  | 16.133126149s |
| hashCode |  2.037175958s | 2.095293416s | 2.011758750s | 1.999053083s | 2.009785958s  | 2.030613433s |
| multiply |  4.233069167s | 4.212958458s | 4.212513125s | 4.230453625s | 4.185100459s  | 4.214818966s |
| parseStringWithMode |  5.962909500s | 5.879412792s | 5.981781583s | 5.899648292s | 5.923697917s  | 5.929490016s |
| pow |  9.133553458s | 9.139799458s | 8.944042083s | 9.104886416s | 9.017410458s  | 9.067938374s |
| roundToDigitPosition |  4.028694209s | 4.146693625s | 3.875029625s | 3.903158042s | 3.845113625s  | 3.959737825s |
| subtract |  5.068563625s | 4.885288833s | 4.796286s | 4.917022458s | 4.777188125s  | 4.888869808s |
| toPlainString |  3.276761500s | 3.318301750s | 3.604505417s | 3.422065917s | 3.557957958s  | 3.435918508s |
| toStringExpanded |  3.206883875s | 3.205201916s | 3.372421625s | 3.204819125s | 3.175107s  | 3.232886708s |


-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- added bench for BigInteger.pow as it's quite often used in BigDecimal and new optimisation was visible


Preparing for BigInteger.pow (warmup)
Benchmarking for BigInteger.pow
Benchmark 0 DONE for BigInteger.pow in 708.251459ms
Benchmark 1 DONE for BigInteger.pow in 591.040042ms
Benchmark 2 DONE for BigInteger.pow in 570.373833ms
Benchmark 3 DONE for BigInteger.pow in 580.99ms
Benchmark 4 DONE for BigInteger.pow in 624.287541ms
Preparing for BigInteger.TEN.pow (warmup)
Benchmarking for BigInteger.TEN.pow
Benchmark 0 DONE for BigInteger.TEN.pow in 197.760459ms
Benchmark 1 DONE for BigInteger.TEN.pow in 172.682916ms
Benchmark 2 DONE for BigInteger.TEN.pow in 171.538708ms
Benchmark 3 DONE for BigInteger.TEN.pow in 181.152167ms
Benchmark 4 DONE for BigInteger.TEN.pow in 174.212041ms
Benchmark report:

| Method |  run 1 | run 2 | run 3 | run 4 | run 5  | Average |
|--------|-------|------|------|------|------ |---------|
| BigInteger.TEN.pow |  197.760459ms | 172.682916ms | 171.538708ms | 181.152167ms | 174.212041ms  | 179.469258ms |
| BigInteger.pow |  708.251459ms | 591.040042ms | 570.373833ms | 580.99ms | 624.287541ms  | 614.988575ms |

…